Nullify batchRequestId immediatelly
Sometimes the library threw unhandledException
that parseParams is undefined in parseResponse.js:312

It happens because a single requestId is assigned to two different
requests, and after the first one is done, the info for second request
is deleted.

A single requestId is assigned because of the batch implementation.
After executeBatch function, the global _batchRequestId is not nulled,
so any further request will get the same requestId, until executeBatch
promise is not finished.